archive/prototype-legacy/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orion-tts"
version = "0.1.0"
description = "Prototype TTS/SRT pipeline toolkit for Orion episodes"
readme = "README.md"
authors = [{name = "Orion Team"}]
license = {text = "Proprietary"}
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "click>=8.1",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3",
]

[project.scripts]
orion-tts = "oriontts.cli:app"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
